Update Alex code to use logging module both in file and console.
