+++ main.py
@@ -1,5 +1,7 @@
 import game
-from game.loggers import logger, file_logger
+from game.loggers import get_logger
+
+logger = get_logger(__name__)
 
 if __name__ == '__main__':
     try:
